DOM Optimization for Minify HTML
Minifying HTML features is like a double-edged sword.
It can change the HTML output and therefore cause some regressions on a given page.
That being said, without DOM optimization, the feature is pretty useless,
especially for servers that use gzip or brotli.

Adding this option would make sense for those who are obsessed with HTML minification
and want to squeeze a few more bytes out of their pages.